Refactor: 상세페이지 데이터 api 통신으로 수정 #19
